Corrected errors to calls to zeros function.
The zeros function is a function provided by the numpy module for
created numpy arrays that are initialized with value 0. However, the
numpy names space was not imported into hmm.py which would have caused
this code to fail upon execution. I've corrected the error by
explictly referencing the numpy namespace when calling the zeros
function.